Split test server into multiple files
