+++ testinit/testing-server/utils.py
@@ -1,0 +1,19 @@
+import logging
+
+# Configure logging
+logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
+logger = logging.getLogger(__name__)
+
+
+def run_cmd(cmd):
+    """Run a shell command and raise if it fails."""
+    logger.info(f"Running command: {cmd}")
+    try:
+        result = subprocess.run(cmd, shell=True, check=True, capture_output=True, text=True)
+        if result.stdout:
+            logger.debug(f"Command output: {result.stdout}")
+        return result
+    except subprocess.CalledProcessError as e:
+        logger.error(f"Command failed: {e}")
+        logger.error(f"Error output: {e.stderr}")
+        raise
